to check later

+++ controllers/contactsControllers.js
@@ -5,22 +5,26 @@ import HttpError from "../helpers/HttpError.js";
 import ctrlWrapper from "../decorators/ctrlWrapper.js";
 
 const getAllContacts = async (req, res) => {
-  const data = await contactsService.listContacts();
+  const { id: owner } = req.user;
+  const data = await contactsService.listContacts({ owner });
   res.json(data);
 };
 
 const getOneContact = async (req, res) => {
   const { id } = req.params;
-  const data = await contactsService.getContactById(id);
+  const { id: owner } = req.user;
+  const data = await contactsService.getContact({ id, owner });
+  console.log(data);
   if (!data) {
-    throw HttpError(404);
+    throw HttpError(404, `Contact with id=${id} not found`);
   }
   res.json(data);
 };
 
 const deleteContact = async (req, res) => {
   const { id } = req.params;
-  const data = await contactsService.removeContact(id);
+  const { id: owner } = req.user;
+  const data = await contactsService.removeContact({ id, owner });
   if (!data) {
     throw HttpError(404);
   }
@@ -28,13 +32,15 @@ const deleteContact = async (req, res) => {
 };
 
 const createContact = async (req, res) => {
-  const data = await contactsService.addContact(req.body);
+  const { id: owner } = req.user;
+  const data = await contactsService.addContact({ ...req.body, owner });
   res.status(201).json(data);
 };
 
 const updateContact = async (req, res) => {
   const { id } = req.params;
-  const data = await contactsService.updateContact(id, req.body);
+  const { id: owner } = req.user;
+  const data = await contactsService.updateContact({ id, owner }, req.body);
 
   if (!data) {
     throw HttpError(404);
